requirements.txt
aiogram>=3.0.0
python-dotenv>=1.0.0
python-docx>=0.8.11
ebooklib>=0.18
beautifulsoup4>=4.12.0
lxml>=4.9.0